feat: refactor file upload
